Move ConditionManager to /api/
